+++ interactive_engine/executor/ir/graph_proxy/src/utils/expr/eval_pred.rs
@@ -17,7 +17,6 @@
 use std::convert::{TryFrom, TryInto};
 
 use dyn_type::{BorrowObject, Object};
-use global_query::store_api::PropId;
 use ir_common::error::ParsePbError;
 use ir_common::expr_parse::error::{ExprError, ExprResult};
 use ir_common::generated::algebra as pb;
@@ -37,19 +36,6 @@ pub struct Predicate {
     pub(crate) left: Operand,
     pub(crate) cmp: common_pb::Logical,
     pub(crate) right: Operand,
-}
-
-impl Predicate {
-    pub(crate) fn extract_prop_ids(&self) -> Option<Vec<PropId>> {
-        let left = self.left.get_var_prop_id();
-        let right = self.right.get_var_prop_id();
-        match (left, right) {
-            (Ok(left), Ok(right)) => Some(vec![left, right]),
-            (Ok(left), _) => Some(vec![left]),
-            (_, Ok(right)) => Some(vec![right]),
-            _ => None,
-        }
-    }
 }
 
 #[allow(dead_code)]
@@ -160,42 +146,6 @@ pub enum Predicates {
     Not(Box<Predicates>),
     And((Box<Predicates>, Box<Predicates>)),
     Or((Box<Predicates>, Box<Predicates>)),
-}
-
-pub(crate) fn zip_option_vecs<T>(left: Option<Vec<T>>, right: Option<Vec<T>>) -> Option<Vec<T>> {
-    match (left, right) {
-        (Some(mut left), Some(mut right)) => {
-            left.append(&mut right);
-            Some(left)
-        }
-        (None, Some(right)) => Some(right),
-        (Some(left), None) => Some(left),
-        (None, None) => None,
-    }
-}
-
-impl Predicates {
-    pub fn extract_prop_ids(&self) -> Option<Vec<PropId>> {
-        match self {
-            Predicates::Init => None,
-            Predicates::SingleItem(operand) => operand
-                .get_var_prop_id()
-                .map(|id| Some(vec![id]))
-                .unwrap_or(None),
-            Predicates::Predicate(pred) => pred.extract_prop_ids(),
-            Predicates::Not(pred) => pred.extract_prop_ids(),
-            Predicates::And((left, right)) => {
-                let left = left.extract_prop_ids();
-                let right = right.extract_prop_ids();
-                zip_option_vecs(left, right)
-            }
-            Predicates::Or((left, right)) => {
-                let left = left.extract_prop_ids();
-                let right = right.extract_prop_ids();
-                zip_option_vecs(left, right)
-            }
-        }
-    }
 }
 
 impl Default for Predicates {
@@ -529,15 +479,6 @@ fn process_predicates(
 pub enum PEvaluator {
     Predicates(Predicates),
     General(Evaluator),
-}
-
-impl PEvaluator {
-    pub fn extract_prop_ids(&self) -> Option<Vec<PropId>> {
-        match self {
-            PEvaluator::Predicates(preds) => preds.extract_prop_ids(),
-            _ => None,
-        }
-    }
 }
 
 impl EvalPred for PEvaluator {
